fixed writing to json

+++ dash_app/charts/burndown_bar/callbacks.py
@@ -29,6 +29,6 @@ def register_callbacks(app):
         df_all = pd.read_json(StringIO(data_all["all"]))
 
         figure, config = processing.create_hours_burndown_chart(
-            df_wertschoepfend, df_all, start_date, end_date, interval, int(wertschoepfend_tage)
+            df_wertschoepfend, df_all, start_date, end_date, interval, float(wertschoepfend_tage)
         )
         return figure, config
